LightSyslogNgConfig: Split DestinationDriver() into separate classes
- new classes: DestinationReader(), DestinationDriver()
- only splitting, functionality is the same
- FileDestination() will use DestinationReader() as a composition
