+++ tests/python_functional/src/syslog_ng_config/statements/destinations/destination_driver.py
@@ -20,36 +20,15 @@
 # COPYING for details.
 #
 #############################################################################
-import logging
-
-from src.message_reader.message_reader import MessageReader
-from src.message_reader.single_line_parser import SingleLineParser
-
-logger = logging.getLogger(__name__)
 
 
 class DestinationDriver(object):
     group_type = "destination"
 
-    def __init__(self, IOClass, positional_parameters=None, options=None):
-        self.__IOClass = IOClass
-        self.__reader = None
+    def __init__(self, positional_parameters=None, options=None):
         if positional_parameters is None:
             positional_parameters = []
         self.positional_parameters = positional_parameters
         if options is None:
             options = {}
         self.options = options
-
-    def dd_read_logs(self, path, counter):
-        if not self.__reader:
-            io = self.__IOClass(path)
-            io.wait_for_creation()
-            message_reader = MessageReader(
-                io.read, SingleLineParser(),
-            )
-            self.__reader = message_reader
-        messages = self.__reader.pop_messages(counter)
-        read_description = "Content has been read from\nresource: {}\ncontent: {}\n".format(path, messages)
-        logger.info(read_description)
-        return messages
